+++ typescript/aleo-sdk/src/clients/protocol.ts
@@ -204,6 +204,7 @@ export class AleoProtocolProvider implements ProtocolProvider {
       TEST_SEND_GAS: 0n,
       AVS_GAS: 0n,
       ISM_DEPLOY_GAS: 0n,
+      HOOK_DEPLOY_GAS: 0n,
     };
   }
 }