+++ typescript/infra/config/environments/testnet4/validators.ts
@@ -247,17 +247,6 @@ export const validatorChainConfig = (
       ),
     },
 
-    incentivtestnet: {
-      interval: 5,
-      reorgPeriod: getReorgPeriod('incentivtestnet'),
-      validators: validatorsConfig(
-        {
-          [Contexts.Hyperlane]: ['0x3133eeb96fd96f9f99291088613edf7401149e6f'],
-        },
-        'incentivtestnet',
-      ),
-    },
-
     radixtestnet: {
       interval: 5,
       reorgPeriod: getReorgPeriod('radixtestnet'),
@@ -288,6 +277,17 @@ export const validatorChainConfig = (
           [Contexts.Hyperlane]: ['0x189ea37efb3764ab638cbaecbe764a8e814e7583'],
         },
         'tronshasta',
+      ),
+    },
+
+    seismictestnet: {
+      interval: 5,
+      reorgPeriod: getReorgPeriod('seismictestnet'),
+      validators: validatorsConfig(
+        {
+          [Contexts.Hyperlane]: ['0xcfa3ec84fe69dde351760e9fdb667af912cb8368'],
+        },
+        'seismictestnet',
       ),
     },
   };